Make both ymir_tools and ymir_common installable
Also refactor utils used by both agents and tools
into the ymir_common package.
