+++ .github/workflows/check-in-container.yml
@@ -20,7 +20,7 @@ jobs:
           - target: check-agents-in-container
           - target: check-mcp-server-in-container
           - target: check-jira-issue-fetcher-in-container
-          - target: check-common-in-container
+          - target: check-ymir-common-in-container
           - target: check-supervisor-in-container
     steps:
     - name: Checkout code
