+++ Containerfile.c9s
@@ -70,7 +70,7 @@ RUN useradd -m -G wheel beeai
 # Individual directories are mounted as volumes in development
 COPY agents/ /home/beeai/agents/
 COPY ymir_tools/ /home/beeai/ymir_tools/
-COPY common/ /home/beeai/common/
+COPY ymir_common/ /home/beeai/ymir_common/
 RUN chgrp -R root /home/beeai && chmod -R g+rwX /home/beeai
 
 # Set up GitLab SSH host keys
