+++ Containerfile.supervisor
@@ -50,7 +50,7 @@ RUN useradd -m -G wheel beeai
 # Individual directories are mounted as volumes in development
 COPY agents/ /home/beeai/agents/
 COPY ymir_tools/ /home/beeai/ymir_tools/
-COPY common/ /home/beeai/common/
+COPY ymir_common/ /home/beeai/ymir_common/
 COPY supervisor/ /home/beeai/supervisor/
 RUN chgrp -R root /home/beeai && chmod -R g+rX /home/beeai
