+++ agents/backport_agent.py
@@ -27,8 +27,8 @@ import tasks
 from agents.build_agent import create_build_agent, get_prompt as get_build_prompt
 from agents.log_agent import create_log_agent, get_prompt as get_log_prompt
 from agents.package_update_steps import PackageUpdateStep, PackageUpdateState
-from common.constants import JiraLabels, RedisQueues
-from common.models import (
+from ymir_common.constants import JiraLabels, RedisQueues
+from ymir_common.models import (
     BackportInputSchema,
     BackportOutputSchema,
     BuildInputSchema,
@@ -39,7 +39,7 @@ from common.models import (
     BackportData,
     ErrorData,
 )
-from common.utils import redis_client, fix_await, is_cs_branch
+from ymir_common.utils import redis_client, fix_await, is_cs_branch
 from constants import I_AM_YMIR, CAREFULLY_REVIEW_CHANGES
 from observability import setup_observability
 from ymir_tools.unprivileged.commands import RunShellCommandTool
@@ -78,7 +78,7 @@ from utils import (
     mcp_tools,
     render_prompt,
 )
-from common.version_utils import is_older_zstream
+from ymir_common.version_utils import is_older_zstream
 from specfile import Specfile
 
 logger = logging.getLogger(__name__)
