+++ agents/rebase_agent.py
@@ -26,9 +26,9 @@ import tasks
 from agents.build_agent import create_build_agent, get_prompt as get_build_prompt
 from agents.log_agent import create_log_agent, get_prompt as get_log_prompt
 from agents.package_update_steps import PackageUpdateStep, PackageUpdateState
-from common.config import get_package_instructions
-from common.constants import JiraLabels, RedisQueues
-from common.models import (
+from ymir_common.config import get_package_instructions
+from ymir_common.constants import JiraLabels, RedisQueues
+from ymir_common.models import (
     BuildInputSchema,
     BuildOutputSchema,
     ErrorData,
@@ -39,7 +39,7 @@ from common.models import (
     RebaseOutputSchema,
     Task,
 )
-from common.utils import redis_client, fix_await, is_cs_branch
+from ymir_common.utils import redis_client, fix_await, is_cs_branch
 from constants import I_AM_YMIR, CAREFULLY_REVIEW_CHANGES
 from observability import setup_observability
 from ymir_tools.unprivileged.commands import RunShellCommandTool
