+++ agents/tasks.py
@@ -8,8 +8,8 @@ from urllib.parse import urlparse
 
 from beeai_framework.tools import Tool
 
-from common.models import LogOutputSchema, CachedMRMetadata, MergeRequestDetails
-from common.utils import is_cs_branch
+from ymir_common.models import LogOutputSchema, CachedMRMetadata, MergeRequestDetails
+from ymir_common.utils import is_cs_branch
 from agents.constants import BRANCH_PREFIX, JIRA_COMMENT_TEMPLATE
 from agents.utils import check_subprocess, run_subprocess, run_tool, mcp_tools
 from ymir_tools.unprivileged.specfile import UpdateReleaseTool
