+++ compose.yaml
@@ -23,7 +23,7 @@ x-beeai-agent: &beeai-agent
   volumes:
     - ./agents:/home/beeai/agents:ro,z
     - ./ymir_tools:/home/beeai/ymir_tools:ro,z
-    - ./common:/home/beeai/common:ro,z
+    - ./ymir_common:/home/beeai/ymir_common:ro,z
     - git-repos:/git-repos:rw,U
     - .secrets/rhel-config.json:/home/beeai/rhel-config.json:ro,z,U
     - .secrets/jotnar-vertex-dev.json:/home/beeai/jotnar-vertex-dev.json:ro,z,U
@@ -69,7 +69,7 @@ x-supervisor: &supervisor
   volumes:
     - ./agents:/home/beeai/agents:ro,z
     - ./ymir_tools:/home/beeai/ymir_tools:ro,z
-    - ./common:/home/beeai/common:ro,z
+    - ./ymir_common:/home/beeai/ymir_common:ro,z
     - ./supervisor:/home/beeai/supervisor:ro,z
     - .secrets/ccache:/home/beeai/ccache:ro,z,U
   restart: unless-stopped
@@ -95,7 +95,7 @@ services:
     volumes:
       - ./ymir_tools/privileged:/home/mcp/ymir_tools/privileged:ro,z
       - ./ymir_tools/privileged/tests/data:/home/mcp/ymir_tools/privileged/tests/data:rw,z
-      - ./common:/home/mcp/common:ro,z
+      - ./ymir_common:/home/mcp/ymir_common:ro,z
       - .secrets/keytab:/home/mcp/keytab:ro,z,U
       - git-repos:/git-repos
       - .secrets/rhel-config.json:/home/mcp/rhel-config.json:ro,z,U
