+++ ymir_common/pyproject.toml
@@ -1,0 +1,26 @@
+[build-system]
+requires = ["setuptools>=45", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "ymir-common"
+version = "0.1.0"
+description = "Common utilities shared across Ymir AI workflows"
+requires-python = ">=3.13"
+dependencies = [
+    "aiofiles>=24.1.0",
+    "beeai-framework[mcp]>=0.1.36",
+    "pydantic>=2.0",
+    "redis>=6.4.0",
+]
+
+[project.optional-dependencies]
+test = [
+    "pytest>=8.4.1",
+    "pytest-asyncio>=1.1.0",
+    "flexmock>=0.12.2",
+]
+
+[tool.setuptools]
+package-dir = {"ymir_common" = "."}
+packages = ["ymir_common"]
