+++ ymir_common/utils.py
@@ -8,11 +8,20 @@ import inspect
 import logging
 import os
 import re
-from contextlib import asynccontextmanager
+import shlex
 import subprocess
-from typing import AsyncGenerator, Awaitable, TypeVar
+from contextlib import asynccontextmanager
+from pathlib import Path
+from typing import Any, AsyncGenerator, Awaitable, Callable, Tuple, TypeVar
 
 import redis.asyncio as redis
+from beeai_framework.middleware.trajectory import GlobalTrajectoryMiddleware
+from beeai_framework.tools import Tool
+from beeai_framework.tools.mcp import MCPTool
+from beeai_framework.tools.types import JSONToolOutput, StringToolOutput
+from mcp import ClientSession
+from mcp.client.sse import sse_client
+from mcp.types import TextContent
 
 logger = logging.getLogger(__name__)
 
@@ -211,3 +220,94 @@ async def init_kerberos_ticket() -> str:
         return principals[0]
     else:
         raise KerberosError("No valid Kerberos ticket found and KEYTAB_FILE is not set")
+
+
+def get_absolute_path(path: Path, tool: Tool) -> Path:
+    if path.is_absolute():
+        return path
+    cwd = (tool.options or {}).get("working_directory") or Path.cwd()
+    return Path(cwd) / path
+
+
+async def run_subprocess(
+    cmd: str | list[str],
+    shell: bool = False,
+    cwd: Path | None = None,
+    env: dict[str, str] | None = None,
+) -> Tuple[int, str | None, str | None]:
+    """Run a subprocess and return the exit code, stdout, and stderr."""
+    kwargs = {
+        "stdout": asyncio.subprocess.PIPE,
+        "stderr": asyncio.subprocess.PIPE,
+    }
+    if cwd is not None:
+        kwargs["cwd"] = cwd
+    if env is not None:
+        kwargs["env"] = os.environ.copy()
+        kwargs["env"].update(env)
+    if shell:
+        if not isinstance(cmd, str):
+            cmd = shlex.join(cmd)
+        proc = await asyncio.create_subprocess_shell(cmd, **kwargs)
+    else:
+        if isinstance(cmd, str):
+            cmd = shlex.split(cmd)
+        proc = await asyncio.create_subprocess_exec(cmd[0], *cmd[1:], **kwargs)
+    stdout, stderr = await proc.communicate()
+    return (
+        proc.returncode,
+        stdout.decode() if stdout else None,
+        stderr.decode() if stderr else None,
+    )
+
+
+async def check_subprocess(
+    cmd: str | list[str],
+    shell: bool = False,
+    cwd: Path | None = None,
+    env: dict[str, str] | None = None,
+) -> Tuple[str | None, str | None]:
+    exit_code, stdout, stderr = await run_subprocess(cmd, shell, cwd, env)
+    if exit_code:
+        logger.error(
+            "Command %s failed with exit code %d\nstdout: %s\nstderr: %s",
+            cmd, exit_code, stdout, stderr,
+        )
+        raise subprocess.CalledProcessError(exit_code, cmd, stdout, stderr)
+    return stdout, stderr
+
+
+async def run_tool(
+    tool: str | Tool,
+    available_tools: list[Tool] | None = None,
+    **kwargs: Any,
+) -> str | dict:
+    if isinstance(tool, str):
+        tool = next(t for t in available_tools or [] if t.name == tool)
+    output = await tool.run(input=kwargs).middleware(GlobalTrajectoryMiddleware(pretty=True))
+    match output:
+        case StringToolOutput():
+            result = output.get_text_content()
+        case JSONToolOutput():
+            result = output.to_json_safe()
+        case _:
+            result = str(output)
+    if isinstance(result, list):
+        [result] = result
+    if isinstance(result, TextContent):
+        result = result.text
+    if isinstance(result, dict) and len(result) == 1 and "result" in result:
+        result = result["result"]
+    return result
+
+
+@asynccontextmanager
+async def mcp_tools(
+    sse_url: str, filter: Callable[[str], bool] | None = None
+) -> AsyncGenerator[list[MCPTool], None]:
+    async with sse_client(sse_url) as (read, write), ClientSession(read, write) as session:
+        await session.initialize()
+        tools = await MCPTool.from_client(session)
+        if filter:
+            tools = [t for t in tools if filter(t.name)]
+        yield tools
