+++ ymir_tools/pyproject.toml
@@ -1,0 +1,35 @@
+[build-system]
+requires = ["setuptools>=45", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "ymir-tools"
+version = "0.1.0"
+description = "Ymir MCP tools for AI workflows"
+requires-python = ">=3.13"
+dependencies = [
+    "ymir-common>=0.1.0",
+    "beeai-framework[mcp]>=0.1.36",
+    "pydantic>=2.0",
+    "aiohttp>=3.12.15",
+    "aiofiles>=24.1.0",
+    "copr>=1.129",
+    "flexmock>=0.12.2",
+    "GitPython>=3.1.0",
+    "koji",
+    "ogr>=0.55.0",
+    "rpm>=0.4.0",
+    "specfile>=0.36.0",
+]
+
+[project.optional-dependencies]
+test = [
+    "pytest>=8.4.1",
+    "pytest-asyncio>=1.1.0",
+    "pytest-cov>=6.0.0",
+    "pytest-mock>=3.14.0",
+]
+
+[tool.setuptools]
+package-dir = {"ymir_tools" = "."}
+packages = ["ymir_tools", "ymir_tools.privileged", "ymir_tools.unprivileged"]
