+++ ymir_tools/unprivileged/filesystem.py
@@ -7,7 +7,7 @@ from beeai_framework.context import RunContext
 from beeai_framework.emitter import Emitter
 from beeai_framework.tools import StringToolOutput, Tool, ToolError, ToolRunOptions
 
-from utils import get_absolute_path
+from ymir_common.utils import get_absolute_path
 
 
 class GetCWDToolInput(BaseModel):
