+++ ymir_tools/unprivileged/specfile.py
@@ -13,9 +13,9 @@ from beeai_framework.context import RunContext
 from beeai_framework.emitter import Emitter
 from beeai_framework.tools import JSONToolOutput, StringToolOutput, Tool, ToolError, ToolRunOptions
 
-from common.constants import BREWHUB_URL
-from common.validators import NonEmptyString
-from agents.utils import get_absolute_path
+from ymir_common.constants import BREWHUB_URL
+from ymir_common.validators import NonEmptyString
+from ymir_common.utils import get_absolute_path
 
 
 class GetPackageInfoToolInput(BaseModel):
