+++ ymir_tools/unprivileged/text.py
@@ -8,8 +8,8 @@ from beeai_framework.context import RunContext
 from beeai_framework.emitter import Emitter
 from beeai_framework.tools import StringToolOutput, Tool, ToolError, ToolRunOptions
 
-from common.validators import NonEmptyString
-from utils import get_absolute_path
+from ymir_common.validators import NonEmptyString
+from ymir_common.utils import get_absolute_path
 
 
 class CreateToolInput(BaseModel):
